refactor: isolate Studio worker lifecycle
