+++ packages/studio/src/clients/worker-live.ts
@@ -32,6 +32,10 @@ import type { FirestoreApi } from '@pyric/ui/firestore';
 import type { AuthApi } from '@pyric/ui/auth';
 import type { StorageApi } from '@pyric/ui/storage';
 import {
+  createStudioWorkerRuntime,
+  type StudioWorkerRuntime,
+} from './worker-runtime.js';
+import {
   getFirestore as workerGetFirestore,
   setOpIssuer,
   getAuth as workerGetAuth,
@@ -45,7 +49,6 @@ import {
   getSnapshot as workerGetSnapshot,
   resetAll as workerResetAll,
   getWorkerInstanceId,
-  getWorkerVersion,
   exportWorkerState,
   importWorkerState,
   saveWorkerBranch,
@@ -87,11 +90,6 @@ import {
   type PresenceSnapshot,
   type PresenceSession,
   readPyricRuntimeManifest,
-  createWorkerReplacement,
-  onWorkerRuntimeReload,
-  preflightWorkerEpochStorage,
-  rememberWorkerEpoch,
-  retireWorkerRuntime,
   workerNameForEpoch,
 } from '@pyric/cli/serve/worker';
 
@@ -105,106 +103,6 @@ interface EpochStorage {
 
 interface RuntimeDocument {
   querySelector(selector: string): { getAttribute(name: string): string | null } | null;
-}
-
-export interface StudioWorkerRuntimeSnapshot {
-  servedEpoch: string | null;
-  runningEpoch: string | null;
-  updateAvailable: boolean;
-  updating: boolean;
-  error: string | null;
-}
-
-export interface StudioWorkerRuntime {
-  getSnapshot(): StudioWorkerRuntimeSnapshot;
-  subscribe(listener: () => void): () => void;
-  update(): Promise<void>;
-}
-
-interface StudioWorkerRuntimeOptions {
-  db: ClientDb;
-  servedEpoch: string | null;
-  storage?: EpochStorage;
-  readVersion?: () => Promise<string>;
-  retire?: (epoch: string) => Promise<void>;
-  subscribeReload?: (listener: (epoch: string) => void) => () => void;
-  preflight?: () => void;
-  remember?: (epoch: string) => void;
-  reload?: () => void;
-  schedule?: (run: () => void) => void;
-}
-
-/** Keep a Studio page on the same explicit worker-replacement lifecycle as the app. */
-export function createStudioWorkerRuntime(
-  options: StudioWorkerRuntimeOptions,
-): StudioWorkerRuntime {
-  const listeners = new Set<() => void>();
-  let snapshot: StudioWorkerRuntimeSnapshot = {
-    servedEpoch: options.servedEpoch,
-    runningEpoch: null,
-    updateAvailable: false,
-    updating: false,
-    error: null,
-  };
-  const publish = (patch: Partial<StudioWorkerRuntimeSnapshot>) => {
-    snapshot = { ...snapshot, ...patch };
-    for (const listener of listeners) listener();
-  };
-  const storage = options.storage;
-  const servedEpoch = options.servedEpoch;
-  const replacement = servedEpoch
-    ? createWorkerReplacement({
-        targetEpoch: servedEpoch,
-        retire: () => options.retire
-          ? options.retire(servedEpoch)
-          : retireWorkerRuntime(options.db, servedEpoch),
-        subscribeReload: options.subscribeReload ?? onWorkerRuntimeReload,
-        preflight: options.preflight ?? (() => preflightWorkerEpochStorage(storage)),
-        commitGeneration: options.remember ?? ((epoch) => rememberWorkerEpoch(epoch, storage)),
-        onPreparationError: (error) => publish({
-          updating: false,
-          error: error instanceof Error ? error.message : String(error),
-        }),
-        reload: options.reload ?? (() => window.location.reload()),
-        ...(options.schedule ? { schedule: options.schedule } : {}),
-      })
-    : null;
-
-  void (options.readVersion ?? (() => getWorkerVersion(options.db)))()
-    .then((runningEpoch) => publish({
-      runningEpoch,
-      updateAvailable: Boolean(
-        servedEpoch
-        && runningEpoch !== 'dev'
-        && runningEpoch !== servedEpoch
-      ),
-    }))
-    .catch((error) => publish({
-      error: error instanceof Error ? error.message : String(error),
-    }));
-
-  return {
-    getSnapshot: () => snapshot,
-    subscribe(listener) {
-      listeners.add(listener);
-      return () => listeners.delete(listener);
-    },
-    async update() {
-      if (!replacement || !snapshot.updateAvailable || snapshot.updating) {
-        throw new Error('No Pyric worker update is available.');
-      }
-      publish({ updating: true, error: null });
-      try {
-        await replacement.request();
-      } catch (error) {
-        publish({
-          updating: false,
-          error: error instanceof Error ? error.message : String(error),
-        });
-        throw error;
-      }
-    },
-  };
 }
 
 /** Resolve the exact worker URL + active generation shared with served apps. */
